requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

pillow==12.2.0
pymupdf==1.27.2

python-multipart==0.0.9
pydantic==2.9.2